# backend/openapi.yaml
openapi: 3.0.0
info:
  title: 'SpotFamServ API'
  description: 'Spotify Familien Server – Backend API (Vertrag, geprüft via oasdiff)'
  version: 1.0.0
servers:
  -
    url: /api/v1
    description: 'API v1'
paths:
...
  /api/v1/auth/me:
    get:
      tags:
        - Auth
      summary: 'Gibt den eingeloggten Admin zurück'
      description: "Liefert Username und Rollen des eingeloggten Admins.\nCatch-all-Regel → 401, wenn keine gültige Session vorhanden."
      operationId: get_api_auth_me
      responses:
        '200':
          description: 'Eingeloggter Admin'
          content:
            application/json:
              schema:
                properties:
                  username: { type: string }
                  roles: { type: array, items: { type: string } }
                type: object
        '401':
          description: 'Nicht eingeloggt'
  /api/v1/auth/login:
    post:
      tags:
        - Auth
      summary: 'Login (json_login; wird von der Firewall abgefangen)'
      description: "Route muss existieren, damit der Router kein 404 wirft.\nDie Firewall (json_login) fängt diese Route ab – dieser Body wird nie erreicht."
      operationId: post_api_auth_login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                username:
                  type: string
                password:
                  type: string
                  format: password
              type: object
      responses:
        '204':
          description: 'Login erfolgreich (Set-Cookie)'
        '401':
          description: 'Ungültige Credentials'
...